// src/generated/models/Dnx_reviewer_commentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Dnx_reviewer_commentsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Dnx_reviewer_commentsstatecode = keyof typeof Dnx_reviewer_commentsstatecode;
export const Dnx_reviewer_commentsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Dnx_reviewer_commentsstatuscode = keyof typeof Dnx_reviewer_commentsstatuscode;

export interface Dnx_reviewer_commentsBase {
  "dnx_Assessment@odata.bind"?: string;
  "dnx_Assessment_Level@odata.bind"?: string;
  dnx_comment_text?: string;
  dnx_is_resolved?: boolean;
  dnx_name: string;
  "dnx_Parent_Comment@odata.bind"?: string;
  dnx_reviewer_commentid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Dnx_reviewer_commentsstatecode;
  statuscode?: Dnx_reviewer_commentsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Dnx_reviewer_comments extends Dnx_reviewer_commentsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  dnx_assessment_levelname?: string;
  dnx_assessmentname?: string;
  dnx_is_resolvedname?: string;
  dnx_parent_commentname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  dnx_assessment?: object;
  _dnx_assessment_value?: string;
  dnx_assessment_level?: object;
  _dnx_assessment_level_value?: string;
  dnx_parent_comment?: object;
  _dnx_parent_comment_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}